+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.139.0] - Not released
+### Fixed
+- The "Download" button in the lightbox was not displayed if the browser did not
+  support the full-screen API.
 
 ## [1.138.1] - 2025-07-09
 ### Fixed